+++ lib/load.php
@@ -40,6 +40,7 @@ if ( ! class_exists( 'WP_Block_Styles_Registry' ) ) {
 require dirname( __FILE__ ) . '/compat.php';
 
 require dirname( __FILE__ ) . '/blocks.php';
+require dirname( __FILE__ ) . '/templates.php';
 require dirname( __FILE__ ) . '/client-assets.php';
 require dirname( __FILE__ ) . '/demo.php';
 require dirname( __FILE__ ) . '/widgets.php';