+++ packages/core-data/src/actions.js
@@ -252,6 +252,8 @@ export function* saveEntityRecord(
 	yield { type: 'SAVE_ENTITY_RECORD_START', kind, name, recordId, isAutosave };
 	let updatedRecord;
 	let error;
+	let persistedEntity;
+	let currentEdits;
 	try {
 		const path = `${ entity.baseURL }${ recordId ? '/' + recordId : '' }`;
 		const persistedRecord = yield select(
@@ -333,6 +335,24 @@ export function* saveEntityRecord(
 			) {
 				data = { ...data, status: 'draft' };
 			}
+
+			// We perform an optimistic update here to clear all the edits that
+			// will be persisted so that if the server filters them, the new
+			// filtered values are always accepted.
+			persistedEntity = yield select(
+				'getEntityRecord',
+				kind,
+				name,
+				recordId
+			);
+			currentEdits = yield select(
+				'getEntityRecordEdits',
+				kind,
+				name,
+				recordId
+			);
+			yield receiveEntityRecords( kind, name, { ...persistedEntity, ...data }, undefined, true );
+
 			updatedRecord = yield apiFetch( {
 				path,
 				method: recordId ? 'PUT' : 'POST',
@@ -342,6 +362,22 @@ export function* saveEntityRecord(
 		}
 	} catch ( _error ) {
 		error = _error;
+
+		// If we got to the point in the try block where we made an optimistic update,
+		// we need to roll it back here.
+		if ( persistedEntity && currentEdits ) {
+			yield receiveEntityRecords( kind, name, persistedEntity, undefined, true );
+			yield editEntityRecord(
+				kind,
+				name,
+				recordId,
+				{
+					...currentEdits,
+					...( yield select( 'getEntityRecordEdits', kind, name, recordId ) ),
+				},
+				{ undoIgnore: true }
+			);
+		}
 	}
 	yield {
 		type: 'SAVE_ENTITY_RECORD_FINISH',